--- src/main/java/domain/Cards.java
@@ -1,38 +1,0 @@
-package domain;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-public class Cards {
-    private final List<Card> cards;
-
-    private Cards(List<Card> cards) {
-        this.cards = cards;
-    }
-
-    public static Cards getDefault() {
-        List<Card> emptyCards = new ArrayList<>();
-        return new Cards(emptyCards);
-    }
-
-    public static Cards of(Card... cards) {
-        List<Card> newCards = new ArrayList<>(Arrays.asList(cards));
-        return new Cards(newCards);
-    }
-
-    public void add(Card card) {
-        cards.add(card);
-    }
-
-    public int countAce() {
-        return (int) cards.stream()
-                .filter(Card::isAce)
-                .count();
-    }
-
-    public List<Card> getCards() {
-        return Collections.unmodifiableList(cards);
-    }
-}
